Switched the z-order of axis to be below series
The axis was changed to be below series to make it possible to get
clicked signals from series when a data point is overlapping an
axis. If the axis is on top of series, you cannot get clicked events
from such data points at all. It should not be a problem if you don't
get a clicked event from the axis in this case; you can always click
on other position on top of the axis (except in the rare use case
that the axis is fully covered by data points of a series).