Refactor the new QXYSeries::pointAt() -> QXYSeries::at()
QML API already had XYSeries::at() for the same functionality.
Also, at() is a common method for returning an item at index in
Qt containers, so use that instead of pointAt().
This function was introduced in polar chart refactoring and has
not yet been released, so no problem changing the name at this point.

Change-Id: I4286178cd971a4f3e6fbff9290ae0bbb5c9097cd
Reviewed-by: Tomi Korpipää <tomi.korpipaa@digia.com>