Fix setting the axis title
The axis title text needs to be set in the ChartAxisElement only if the
new title is empty or the title is hidden. In other cases the title will
be set in updateGeometry() of the axis.

Removed redundancy of setting title information. The title information
needs to be set only once when creating items for axis.

Task-number: QTRD-2104
Change-Id: Ie29c94215bdbab412a3dbb8368309abbd15aa4d1
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@digia.com>