Fix debug-and-release build
Charts is part of the Qt release and there is no need to do magic hacks
to install the qmldir file.
This was supposed to help running examples separately, but we assume
that the lib is shipped with Qt proper.

The problem was that the qmldir file would be simultaneously copied into
the same location twice which will fail on Windows.

Change-Id: I4d765c18367f7985c7dbfadbc7ddae78ce24f398
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@theqtcompany.com>
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@theqtcompany.com>
Reviewed-by: Titta Heikkala <titta.heikkala@theqtcompany.com>