Ignore mouse events in pie series
In case of overlapping pie series, the topmost series was eating
mouse events even when there were no slices belonging to that series
under the mouse. Fixed by explicitly ignoring all mouse events that
are offered for the main series item (PieChartItem).
This is okay, since we are only interested in events from the pie
slices.

Task-number: QTBUG-57178
Change-Id: I2e7faefce722a3685cdb71a8e12654f408c73a51
Reviewed-by: Mika Salmela <mika.salmela@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>