Maintain correct order of legend items when inserting pie chart items
Rewrite QLegendPrivate::handleCountChanged() so that the order
of the markers is preserved. Split out helper functions for
insertion/removal that leave the m_markers list untouched and
add helpers for finding markers by series and related objects.
In QLegendPrivate::handleCountChanged(), remove the markers
of the series from m_markers and replace the existing ones
in the createdMarkers. Finally re-insert createdMarkers into
m_markers, preserving the order.

Task-number: QTBUG-62082
Change-Id: I03c81dcf5dfd5a5883377630cad16c943503a1be
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>