Fix data normalization in GLXYSeriesDataManager
GLXYSeriesDataManager::setPoints() has an optimized path when it detects
a regular XY chart, which skips calculateGeometryPoints() and sets up
a shader-based transform instead of normalizing vertex data. However,
for some series data containing values far away from the origin, which are
represented with less absolute precision as floating-point values, by the
time the transform is used, precision was already lost in the conversion
from double to float, destroying the chart. This change performs proper
normalization inside the optimized path, while keeping the same level of
performance as before.

Task-number: QTBUG-59674
Change-Id: Idc57481815a409c19ca364e1721153483cbc2993
Reviewed-by: Andy Shaw <andy.shaw@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>