Add label editing feature to the QValueAxis
Add a new property (labelsEditable) to the QAbstractAxis. If this
property set to true the user will be able to manipulate the axis
labels by double clicking on it and entering a new value. After
finishing the edit the axis range will adjusted according to the
entered value. At the moment only the QValueAxis is supported,
but I am planning to add support for QLogValueAxis, and QDateTimeAxis
after the basic implementation details got reviewed.

Change-Id: Id87ebfbfb37ed00f09cf7b651395cf3caead1dbb
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>