Fully qualify classes in QtCharts namespace in Q_PROPERTY
qmlplugindump otherwise cannot make the connection between the
type of the property and the QMetaObject. This fixes an issue where
Qt Creator was not handling some properties for the QML types
properly.

Ideally we also would do this for arguments in signals, slots,
Q_INVOKABLES. Anyhow, this can be an incompatible change (e.g.
if you use string-based connect).

Fixes: QTBUG-51704
Change-Id: I3a3b5fc7f7539ff15974be9c2e268df7e1ad7f30
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>