Fix rendering lines with Antialiasing render hint
When rendering with QPainter::Antialiasing set, lines need to be drawn
in one set. Otherwise, each line is rendered with the anti-aliasing
applied, causing the overlap to create a stronger area, which almost
looks like a point from the point series. As a consequence, prefer
drawPath() in that case.

Task-number: QTBUG-61782
Change-Id: Ib2b44475e2e9ca0c28da7e898744bd72b992f5fd
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>