QtCharts: Append branch condition fix in QLegend
Qlist insert was called with -1 as position in
QLegendPrivate::handleCountChanged, as pos = indexOfSeries(series,
m_markers) is used as insert pos, which has -1 if the series is not
found in the markers. The branch condition that should handle this use
case was mistaken and the else branch executed which was not inteded.
The else branch was still functional since the insert position
protection adjusted it to 0 but it gave an assert because of the -1 as
insert position. The assert and warning for negative indices was added
in 5.15, which is why there was no warning before.

Corrected the condition. The issue doesn't look platform dependent.

Fixes: QTBUG-85909
Change-Id: I78c7dba3b4793ad23f1508c11bd43bca8b4533ed
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>