Add custom legend markers to XYChart based charts
This is a feature which uses the light marker assigned to a series as
the marker in the legend. The Scatter chart had similar functionality,
but the Line and Spline charts did not. The Scatter chart is still able
to use the series brush as a marker as previously implemented. This
change does not break backward compatibility with Scatter charts.

The marker size property of the Scatter series was moved to XYSeries so
it can be used by Line and Spline. This has the benefit of allowing the
user to specify the size of the light marker on the chart and on the
legend. The new API allowed the legend to deal with any XYSeries
subclass without regard to its implementation.

Instead of adding a new QGraphicsViewItem to the LegendMarkerItem, the
light marker, if present, is simply painted on the LegendMarkerItem.
This was the simplest path to implement the desired feature and did not
require a complete rework of the existing marker item code.

This also adds rotated rectangle, triangle, star and pentagon to the
explicitly set shape types that can be specified in the legend.

Task-number: QTBUG-93365
Change-Id: I2ee3cbbe9eae051f48d4528fbde351caf5e2922f
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>