Extract precision-needed calculation to a function
Two of the three places doing the calculation neglected to consider
the case of only one tick (division by ticks - 1 is then a problem)
and only one of these explained what the calculation was about.

Package it in a little function and explain what it's about a little
more clearly, then call it from all three users and avoid repetition.

Change-Id: I5826e37522a785d50ba103976be33f4c06ce1dea
Reviewed-by: Dimitrios Apostolou <jimis@qt.io>
Reviewed-by: Michal Klocek <michal.klocek@qt.io>