Use ratios of natural logs rather than to base ten
Natural log is marginally cheaper to compute and a ratio of logs, to a
given base, doesn't actually depend on the base; log(a)/log(b) is the
logarithm of a to base b.

Tidy up some duplication into a use of a ternary expression to
simplify one use of this, in the process.

Change-Id: Ic91dc0b4e9255f20f2f6e3d71c32a37f759d623d
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>