Divide by logBase in the right place to save messy computations
All logs used in computing layout should be to the axis's base;
natural logs should only be used to compute those. This incidentally
saves the "need" to raise the base to a power in order to then take
the log of the result (which would, in any case, have been better done
by just multiplying logBase by the power involved).

The finally computed values do in fact come out the same, as the
factors of logBase are merely moved around and end up cancelling in
computing the extra tick values, but using logs to the right base
makes the code easier to understand.

Change-Id: I6c6921f08ea8359d8236db15ef16eb8ddedb3d29
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>